+++ coffeenet-autoconfigure/src/test/java/coffee/synyx/autoconfigure/navigation/CoffeeNetNavigationInterceptorTest.java
@@ -24,7 +24,7 @@ import static org.mockito.Mockito.when;
  * @author  Tobias Schneider
  */
 @RunWith(MockitoJUnitRunner.class)
-public class CoffeeNetNavigationInformationInterceptorTest {
+public class CoffeeNetNavigationInterceptorTest {
 
     private CoffeeNetNavigationInterceptor sut;